Turn status plugin into folder
Remove the use of an override and add a hook `presenceConstructed` instead
